requirements.txt
langchain
openai
requests
langchain_community
langchain-openai
python-dotenv